requirements.txt
streamlit>=1.38.0
numpy
pandas>=2.1.0
matplotlib
plotly
seaborn
altair
google-cloud-bigquery
gspread
langchain
langchain-experimental
openai
chromadb>=0.4.0
tiktoken
pygwalker
pytrends
flatten_json
streamlit_authenticator
